+++ src/solutions/year2021/day02.rs
@@ -1,13 +1,13 @@
 use crate::grid::{E, N, S, Z};
 
-fn parse(input: &str) -> impl Iterator<Item = (&str, i32)> + '_ {
+fn parse(input: &str) -> impl Iterator<Item = (&str, i64)> + '_ {
     input.trim().lines().map(|line| {
         let (dir, n) = line.trim().split_once(' ').unwrap();
         (dir, n.parse().unwrap())
     })
 }
 
-pub fn part1(input: &str) -> i32 {
+pub fn part1(input: &str) -> i64 {
     let mut pos = Z;
     for (dir, n) in parse(input) {
         match dir {
@@ -20,7 +20,7 @@ pub fn part1(input: &str) -> i32 {
     pos.x * pos.y
 }
 
-pub fn part2(input: &str) -> i32 {
+pub fn part2(input: &str) -> i64 {
     let mut pos = Z;
     let mut aim = E;
     for (dir, n) in parse(input) {
